Replace GORM, lots of testing, rewrite GRPC call semantics
This change overall:

- Replaces GORM with SQLBoiler. This has already resulted in some pretty
  impressive performance improvements, but the code is also much simpler
  too.
  - model <-> proto conversion routines were redone
    - see under GRPC for more consequences here
- Rewrites the db test subsystem to use gopkg.in/assert
  - This library is much easier to use and gives reliable output
  - Other suites were not ported, that will come later
- GRPC Call semantics were changed
  - the grpc types library is leaned on harder now and must translate to
    json reliably in a few spots
  - the model types are gone
  - topTypes aka ci-agents/types (vs ci-agents/ci-gen/grpc/types) is now
    used to represent settings and other non-structural data, but
    otherwise the code relies on the above protobuf conversion code to
    translate between the db and back.

+++ utils/pagination.go
@@ -17,7 +17,8 @@ func intp(p int64) *int64 {
 
 // ScopePaginationInt applies constraints to pagination; it sets a maximum and a
 // default if not supplied to perPage. Integer-only version.
-func ScopePaginationInt(page, perPage *int64) (int64, int64, error) {
+func ScopePaginationInt(page, perPage *int64) (int, int, error) {
+	// FIXME make this int32 everywhere
 	if page == nil {
 		page = intp(0)
 	}
@@ -42,12 +43,12 @@ func ScopePaginationInt(page, perPage *int64) (int64, int64, error) {
 		return 0, 0, errors.New("invalid per page")
 	}
 
-	return *page, *perPage, nil
+	return int(*page), int(*perPage), nil
 }
 
 // ScopePagination applies constraints to pagination; it sets a maximum and a
 // default if not supplied to perPage. String version.
-func ScopePagination(pg, ppg string) (int64, int64, error) {
+func ScopePagination(pg, ppg string) (int, int, error) {
 	var (
 		page, perPage int64
 	)
